requirements.txt
aiohttp
aiosqlite
discord.py
fastapi
psutil
pydantic
python-dotenv
uvicorn
